requirements.txt
altair==4.2.0
appdirs==1.4.4
attrs==22.1.0
backoff==2.2.1
black==22.10.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
decorator==5.1.1
Deprecated==1.2.13
entrypoints==0.4
gitdb==4.0.10
GitPython==3.1.29
gql==3.4.0
graphql-core==3.2.3
idna==3.4
importlib-metadata==5.1.0
Jinja2==3.1.2
jsonschema==4.17.1
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.23.5
packaging==21.3
pandas==1.5.2
pathspec==0.10.2
patsy==0.5.3
Pillow==9.3.0
platformdirs==2.5.4
plotly==5.11.0
plotly-express==0.4.1
protobuf==3.20.3
pyarrow==10.0.1
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
scipy==1.9.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
specklepy==2.9.1
statsmodels==0.13.5
streamlit==1.15.1
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.13
validators==0.20.0
watchdog==2.1.9
websockets==10.4
wrapt==1.14.1
yarl==1.8.1
zipp==3.11.0
